chore: exception handling

+++ src/main/java/dlanaras/com/github/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+import javax.persistence.EntityExistsException;
 import javax.persistence.EntityManager;
+import javax.persistence.TransactionRequiredException;
 import javax.transaction.Transactional;
 
 import dlanaras.com.github.exceptions.InvalidLoginException;
+import dlanaras.com.github.exceptions.NullValueException;
 import dlanaras.com.github.models.User;
 import dlanaras.com.github.models.dto.Login;
 
@@ -22,18 +25,34 @@ public class UserService {
     }
 
     @Transactional
-    public User createUser(User user) {
-        entityManager.persist(user);
+    public User createUser(User user) throws EntityExistsException, Exception {
+        try {
+            entityManager.persist(user);
+        } catch (EntityExistsException e) {
+            throw e;
+        } catch (Exception e) {
+            throw e;
+        }
+
         return user;
     }
 
     @Transactional
-    public User updateUser(User user) {
-        return entityManager.merge(user);
+    public User updateUser(User user) throws IllegalArgumentException, TransactionRequiredException {
+        try {
+            return entityManager.merge(user);
+        } catch (IllegalArgumentException e) {
+            throw e;
+        } catch (TransactionRequiredException e) {
+            throw e;
+        }
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws NullValueException {
         var entity = entityManager.find(User.class, id);
+        if (entity == null) {
+            throw new NullValueException("No user with id: " + id + " was found");
+        }
         entityManager.remove(entity);
     }
 
@@ -49,8 +68,14 @@ public class UserService {
         return "THIS IS A TOKEN";
     }
     
-    public String register(User user) {
-        createUser(user);
+    public String register(User user) throws EntityExistsException, Exception {
+        try {
+            createUser(user);
+        } catch (EntityExistsException e) {
+            throw e;
+        } catch (Exception e) {
+            throw e;
+        }
 
         //TODO: return real token
         return "This is a token";
